refacotr : 관리자 페이지 제거
