--- src/components/common/index.ts
@@ -1,7 +1,0 @@
-export { default as StatisticsCard } from './StatisticsCard';
-export { default as SearchBar } from './SearchBar';
-export { default as FilterDropdown } from './FilterDropdown';
-export { default as DataTable } from './DataTable';
-export { default as ActionButton } from './ActionButton';
-export { default as Pagination } from './Pagination';
-export { default as RankingList } from './RankingList';
