+++ src/features/allBenefitsPage/index.tsx
@@ -3,10 +3,10 @@ import EventBanner from './components/EventBanner';
 import SimpleRanking from './components/SimpleRanking';
 import { useState, useEffect, useRef, useCallback, useMemo } from 'react';
 import { debounce } from 'lodash';
-import BenefitFilterToggle from '../../components/common/BenefitFilterToggle';
-import SearchBar from '../../components/common/SearchBar';
-import Pagination from '../../components/common/Pagination';
-import LoadingSpinner from '../../components/common/LoadingSpinner';
+import BenefitFilterToggle from '../../components/BenefitFilterToggle';
+import SearchBar from '../../components/SearchBar';
+import Pagination from '../../components/Pagination';
+import LoadingSpinner from '../../components/LoadingSpinner';
 import NoResult from '../../components/NoResult';
 import { TbChevronDown, TbStar, TbStarFilled } from 'react-icons/tb';
 import { showToast } from '../../utils/toast';
