--- src/pages/adminPage/AdminPage.tsx
@@ -1,6 +1,0 @@
-import AdminLayout from '../../features/adminPage/components/Layout';
-const AdminPage = () => {
-  return <AdminLayout />;
-};
-
-export default AdminPage;
